+++ src/pymmcore_widgets/__init__.py
@@ -8,25 +8,28 @@ try:
 except PackageNotFoundError:
     __version__ = "uninstalled"
 
-from ._camera_roi_widget import CameraRoiWidget
-from ._channel_group_widget import ChannelGroupWidget
-from ._channel_widget import ChannelWidget
-from ._device_widget import DeviceWidget, StateDeviceWidget
-from ._exposure_widget import DefaultCameraExposureWidget, ExposureWidget
-from ._group_preset_widget._group_preset_table_widget import GroupPresetTableWidget
-from ._image_widget import ImagePreview
+from ._deprecated._device_widget import DeviceWidget, StateDeviceWidget
 from ._install_widget import InstallWidget
-from ._live_button_widget import LiveButton
-from ._load_system_cfg_widget import ConfigurationWidget
-from ._objective_widget import ObjectivesWidget
-from ._objectives_pixel_configuration_widget import ObjectivesPixelConfigurationWidget
-from ._pixel_configuration_widget import PixelConfigurationWidget
-from ._presets_widget import PresetsWidget
-from ._properties_widget import PropertiesWidget
-from ._property_browser import PropertyBrowser
-from ._shutter_widget import ShuttersWidget
-from ._snap_button_widget import SnapButton
-from ._stage_widget import StageWidget
+from .config_presets._group_preset_widget._group_preset_table_widget import (
+    GroupPresetTableWidget,
+)
+from .config_presets._objectives_pixel_configuration_widget import (
+    ObjectivesPixelConfigurationWidget,
+)
+from .config_presets._pixel_configuration_widget import PixelConfigurationWidget
+from .control._camera_roi_widget import CameraRoiWidget
+from .control._channel_group_widget import ChannelGroupWidget
+from .control._channel_widget import ChannelWidget
+from .control._exposure_widget import DefaultCameraExposureWidget, ExposureWidget
+from .control._live_button_widget import LiveButton
+from .control._load_system_cfg_widget import ConfigurationWidget
+from .control._objective_widget import ObjectivesWidget
+from .control._presets_widget import PresetsWidget
+from .control._shutter_widget import ShuttersWidget
+from .control._snap_button_widget import SnapButton
+from .control._stage_widget import StageWidget
+from .device_properties._properties_widget import PropertiesWidget
+from .device_properties._property_browser import PropertyBrowser
 from .device_properties._property_widget import PropertyWidget
 from .hcwizard import ConfigWizard
 from .mda import MDAWidget
@@ -38,6 +41,7 @@ from .useq_widgets import (
     TimePlanWidget,
     ZPlanWidget,
 )
+from .views._image_widget import ImagePreview
 
 
 def __getattr__(name: str) -> object:
