+++ src/pymmcore_widgets/config_presets/_group_preset_widget/__init__.py
@@ -1,0 +1,15 @@
+from ._add_first_preset_widget import AddFirstPresetWidget
+from ._add_group_widget import AddGroupWidget
+from ._add_preset_widget import AddPresetWidget
+from ._edit_group_widget import EditGroupWidget
+from ._edit_preset_widget import EditPresetWidget
+from ._group_preset_table_widget import GroupPresetTableWidget
+
+__all__ = [
+    "AddGroupWidget",
+    "AddFirstPresetWidget",
+    "AddPresetWidget",
+    "EditGroupWidget",
+    "EditPresetWidget",
+    "GroupPresetTableWidget",
+]
