+++ src/pymmcore_widgets/control/__init__.py
@@ -1,0 +1,1 @@
+"""Widgets that control various devices at runtime."""
