+++ src/pymmcore_widgets/control/_camera_roi_widget.py
