+++ src/pymmcore_widgets/control/_channel_group_widget.py
