+++ src/pymmcore_widgets/control/_channel_widget.py
@@ -3,8 +3,8 @@ from __future__ import annotations
 from pymmcore_plus import CMMCorePlus
 from qtpy.QtWidgets import QComboBox, QVBoxLayout, QWidget
 
-from ._presets_widget import PresetsWidget
-from ._util import ComboMessageBox
+from pymmcore_widgets._util import ComboMessageBox
+from pymmcore_widgets.control._presets_widget import PresetsWidget
 
 
 class ChannelWidget(QWidget):
