+++ src/pymmcore_widgets/control/_exposure_widget.py
