+++ src/pymmcore_widgets/control/_shutter_widget.py
