+++ src/pymmcore_widgets/control/_snap_button_widget.py
