+++ src/pymmcore_widgets/control/_stage_widget.py
