+++ src/pymmcore_widgets/device_properties/_device_property_table.py
@@ -3,36 +3,19 @@ from __future__ import annotations
 from logging import getLogger
 from typing import Iterable, cast
 
-from fonticon_mdi6 import MDI6
 from pymmcore_plus import CMMCorePlus, DeviceProperty, DeviceType
 from qtpy.QtCore import Qt
 from qtpy.QtGui import QColor
 from qtpy.QtWidgets import QAbstractScrollArea, QTableWidget, QTableWidgetItem, QWidget
 from superqt.fonticon import icon
 
-from pymmcore_widgets.device_properties._property_widget import PropertyWidget
+from pymmcore_widgets._icons import ICONS
 
-ICONS: dict[DeviceType, str] = {
-    DeviceType.Any: MDI6.devices,
-    DeviceType.AutoFocus: MDI6.auto_upload,
-    DeviceType.Camera: MDI6.camera,
-    DeviceType.Core: MDI6.checkbox_blank_circle_outline,
-    DeviceType.Galvo: MDI6.mirror_variant,
-    DeviceType.Generic: MDI6.dev_to,
-    DeviceType.Hub: MDI6.hubspot,
-    DeviceType.ImageProcessor: MDI6.image_auto_adjust,
-    DeviceType.Magnifier: MDI6.magnify_plus,
-    DeviceType.Shutter: MDI6.camera_iris,
-    DeviceType.SignalIO: MDI6.signal,
-    DeviceType.SLM: MDI6.view_comfy,
-    DeviceType.Stage: MDI6.arrow_up_down,
-    DeviceType.State: MDI6.state_machine,
-    DeviceType.Unknown: MDI6.dev_to,
-    DeviceType.XYStage: MDI6.arrow_all,
-    DeviceType.Serial: MDI6.serial_port,
-}
+from ._property_widget import PropertyWidget
 
 logger = getLogger(__name__)
+
+__all__ = ["DevicePropertyTable"]
 
 
 class DevicePropertyTable(QTableWidget):
