+++ src/pymmcore_widgets/experimental.py
@@ -1,3 +1,3 @@
-from ._stack_viewer import StackViewer
+from .views._stack_viewer import StackViewer
 
 __all__ = ["StackViewer"]
