+++ src/pymmcore_widgets/views/__init__.py
@@ -1,0 +1,1 @@
+"""Widgets that view images and other states."""
