+++ src/pymmcore_widgets/views/_stack_viewer/_channel_row.py
