+++ src/pymmcore_widgets/views/_stack_viewer/_labeled_slider.py
